feat 오류 수정 및 함수분리
