+++ src/app/(providers)/layout.tsx
@@ -1,14 +1,11 @@
 import QueryProvider from '@/providers/QueryProvider';
 import SupabaseProvider from '@/providers/SupabaseProvider';
-import { CounterStoreProvider } from '@/providers/CounterStoreProvider';
 import type { StrictPropsWithChildren } from '@/types/common';
 
 const ProvidersLayout = ({ children }: StrictPropsWithChildren) => {
   return (
     <QueryProvider>
-      <SupabaseProvider>
-        <CounterStoreProvider>{children}</CounterStoreProvider>
-      </SupabaseProvider>
+      <SupabaseProvider>{children}</SupabaseProvider>
     </QueryProvider>
   );
 };
