--- src/providers/CounterStoreProvider.tsx
@@ -1,34 +1,0 @@
-'use client';
-
-import { useStore } from 'zustand';
-import { type ReactNode, createContext, useRef, useContext } from 'react';
-import { type CounterStore, createCounterStore, initCounterStore } from '@/stores/counterStore';
-
-export interface CounterStoreProviderProps {
-  children: ReactNode;
-}
-export type CounterStoreApi = ReturnType<typeof createCounterStore>;
-
-export const CounterStoreContext = createContext<CounterStoreApi | undefined>(undefined);
-
-export const CounterStoreProvider = ({ children }: CounterStoreProviderProps) => {
-  const storeRef = useRef<CounterStoreApi>();
-
-  if (!storeRef.current) {
-    storeRef.current = createCounterStore(initCounterStore());
-  }
-
-  return (
-    <CounterStoreContext.Provider value={storeRef.current}>{children}</CounterStoreContext.Provider>
-  );
-};
-
-export const useCounterStore = <T,>(selector: (store: CounterStore) => T): T => {
-  const counterStoreContext = useContext(CounterStoreContext);
-
-  if (!counterStoreContext) {
-    throw new Error(`useCounterStore must be used within CounterStoreProvider`);
-  }
-
-  return useStore(counterStoreContext, selector);
-};
