--- src/stores/counterStore.ts
@@ -1,28 +1,0 @@
-import { createStore } from 'zustand/vanilla';
-
-export type CounterState = {
-  count: number;
-};
-
-export type CounterActions = {
-  decrementCount: () => void;
-  incrementCount: () => void;
-};
-
-export type CounterStore = CounterState & CounterActions;
-
-export const initCounterStore = (): CounterState => {
-  return { count: new Date().getFullYear() };
-};
-
-export const defaultInitState: CounterState = {
-  count: 0
-};
-
-export const createCounterStore = (initState: CounterState = defaultInitState) => {
-  return createStore<CounterStore>()((set) => ({
-    ...initState,
-    decrementCount: () => set((state) => ({ count: state.count - 1 })),
-    incrementCount: () => set((state) => ({ count: state.count + 1 }))
-  }));
-};
