--- src/stores/userStores.ts
@@ -1,14 +1,0 @@
-import { User } from '@supabase/supabase-js';
-import create from 'zustand';
-
-export type Store = {
-  user: User | null;
-  setUser: (user: User | null) => void;
-  clearUser: () => void;
-};
-
-export const useUserStore = create<Store>((set) => ({
-  user: null,
-  setUser: (user) => set({ user }),
-  clearUser: () => set({ user: null })
-}));
